+++ client-app/modules/push-messages/api/graphql/fragments/pushMessage.graphql
