+++ client-app/modules/push-messages/api/graphql/mutations/clearAllPushMessages/clearAllPushMessagesMutation.graphql
