+++ client-app/modules/push-messages/api/graphql/mutations/clearAllPushMessages/index.ts
