+++ client-app/modules/push-messages/api/graphql/mutations/deleteFcmToken/deleteFcmToken.graphql
