+++ client-app/modules/push-messages/api/graphql/mutations/markAllPushMessagesRead/index.ts
