+++ client-app/modules/push-messages/api/graphql/mutations/markAllPushMessagesRead/markAllPushMessagesReadMutation.graphql
