+++ client-app/modules/push-messages/api/graphql/mutations/markAllPushMessagesUnread/markAllPushMessagesUnread.graphql
