+++ client-app/modules/push-messages/api/graphql/mutations/markPushMessageRead/markPushMessageRead.graphql
