+++ client-app/modules/push-messages/api/graphql/queries/getPushMessages/onPushMessageCreatedSubscription.graphql
