+++ client-app/router/routes/account.ts
@@ -1,5 +1,5 @@
 import { useUser } from "@/shared/account";
-import { isActive as isPushMessagesActive } from "@/shared/push-messages/composables/usePushMessages";
+import { isActive as isPushMessagesActive } from "@/modules/push-messages/composables/usePushMessages";
 import type { RouteRecordRaw } from "vue-router";
 
 const Dashboard = () => import("@/pages/account/dashboard.vue");
@@ -13,7 +13,7 @@ const Lists = () => import("@/pages/account/lists.vue");
 const ListDetails = () => import("@/pages/account/list-details.vue");
 const CheckoutDefaults = () => import("@/pages/account/checkout-defaults.vue");
 const SavedCreditCards = () => import("@/pages/account/saved-credit-cards.vue");
-const Notifications = () => import("@/pages/account/notifications.vue");
+const Notifications = () => import("@/modules/push-messages/pages/notifications.vue");
 const Impersonate = () => import("@/pages/account/impersonate.vue");
 
 export const accountRoutes: RouteRecordRaw[] = [
